// extensions/discord/src/send.types.ts
export type DiscordRuntimeAccountContext = {
  cfg: ACTAgentConfig;
  accountId: string;
};

export type DiscordReactOpts = {
  cfg: ACTAgentConfig;
  accountId?: string;
  token?: string;
  rest?: RequestClient;
  verbose?: boolean;
  retry?: RetryConfig;
};

export type DiscordReactionRuntimeContext = DiscordRuntimeAccountContext & {
  rest: RequestClient;
};